Add click to activate option
To keep the default behavior of clicking on a workspace indicator more in line with what people might expect of an indicator button the option to click to activate a workspace has been added, with the default action on either left or right click to be to open the menu.

Also added is the option to specify whether the Primary or Secondary button will activate the workspace, allowing for preference.

Various bits of code within the preferences are also been tidied.