Fix object access errors
Using the Mainloop delay within the _updateWhatever type functions appears to cause issues with object access (accessing deallocated and finalized objects), triggered more readily on Wayland. This removes the use of the Mainloop and simply uses a return false to stop execution inside the functions, since whatever was returning null (hence the use of the Mainloop) probably is showing that the object doesn't exist and we trigger updates often enough that we should end up updating whatever it is we needed to in the end.